manually_instrumented_fastapi_service/pyproject.toml
[tool.poetry]
name = "fastapiservice"
version = "0.1.0"
description = ""
authors = ["patex1987 <patex1987@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
opentelemetry-sdk = "^1.28.2"
opentelemetry-api = "^1.28.2"
opentelemetry-distro = "^0.49b2"
opentelemetry-exporter-otlp = "^1.28.2"
opentelemetry-instrumentation-fastapi = "^0.49b2"
requests = "^2.32.3"
opentelemetry-instrumentation-requests = "0.49b2"
httpx = "^0.28.1"
opentelemetry-instrumentation-httpx = "0.49b2"
structlog = "^25.1.0"
svcs = "^25.1.0"
pydantic = "^2.10.6"


[tool.poetry.group.dev.dependencies]
ipython = "^8.32.0"
pudb = "^2024.1.3"
debugpy = "^1.8.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
